move existing code into package
